Fix formatting of qualifying times resolves #8
