+++ tests/fixtures/ergast_responses/laps/2023@1@drivers@leclerc@qualifying.json
@@ -1,0 +1,1 @@
+{"MRData":{"xmlns":"http:\/\/ergast.com\/mrd\/1.5","series":"f1","url":"http://ergast.com/api/f1/2023/1/drivers/leclerc/qualifying.json","limit":"30","offset":"0","total":"1","RaceTable":{"season":"2023","round":"1","driverId":"leclerc","Races":[{"season":"2023","round":"1","url":"https://en.wikipedia.org/wiki/2023_Bahrain_Grand_Prix","raceName":"Bahrain Grand Prix","Circuit":{"circuitId":"bahrain","url":"http://en.wikipedia.org/wiki/Bahrain_International_Circuit","circuitName":"Bahrain International Circuit","Location":{"lat":"26.0325","long":"50.5106","locality":"Sakhir","country":"Bahrain"}},"date":"2023-03-05","time":"15:00:00Z","QualifyingResults":[{"number":"16","position":"3","Driver":{"driverId":"leclerc","permanentNumber":"16","code":"LEC","url":"http://en.wikipedia.org/wiki/Charles_Leclerc","givenName":"Charles","familyName":"Leclerc","dateOfBirth":"1997-10-16","nationality":"Monegasque"},"Constructor":{"constructorId":"ferrari","url":"http://en.wikipedia.org/wiki/Scuderia_Ferrari","name":"Ferrari","nationality":"Italian"},"Q1":"1:31.094","Q2":"1:30.282","Q3":"1:30.000"}]}]}}}
